+++ src/main/java/com/raoulvdberge/refinedstorage/apiimpl/network/node/NetworkNodeStorage.java
@@ -323,5 +323,9 @@ public class NetworkNodeStorage extends NetworkNode implements IGuiStorage, ISto
         this.priority = priority;
 
         markDirty();
+
+        if (network != null) {
+            network.getItemStorageCache().sort();
+        }
     }
 }